+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 * **根目录:** 在项目中配置conventional-changelog 8814a8e
 * **dllplugin:** 使用DLLPlugin 提取公共包，使用DLLReferencePlugin引用公共包的映射,但是还要再html文件中手动引入公共包的js文件。 d8512d7
+* **webpack.prod.js:** 缓存：开启babel缓存提升打包速度。 hard-source-webpack-plugin由于webpack版本问题添加失败 e248500
 * 并行压缩 c34bc45